refactor: fix icon import
