+++ components/Shared/Svg/Icons/index.ts
@@ -1,1 +1,1 @@
-export { default as ArrowDown } from './ArrowDown';
+export { default as ArrowDownIcon } from './ArrowDownIcon';
